PDA入库抽样显示该批次到货总数

+++ wms-cloud/wms-api/storage-in-api/src/main/java/com/bosch/storagein/api/domain/dto/MaterialInCheckDTO.java
@@ -12,7 +12,7 @@ public class MaterialInCheckDTO {
     private String mesBarCode;
 
     @ApiModelProperty(value = "实际数量")
-    private Integer actualQuantity;
+    private Double actualQuantity;
 
     @ApiModelProperty(value = "实际称重、数数结果")
     private Double actualResult;
